TAF-14 Moving utility classes, models, exceptions to test-common
