+++ skyxplore-ui-test/src/test/java/com/mentoring/pageobject/LobbyQueuePage.java
@@ -1,6 +1,6 @@
 package com.mentoring.pageobject;
 
-import com.mentoring.framework.Config;
+import com.mentoring.config.Config;
 import org.openqa.selenium.WebDriver;
 
 import java.util.concurrent.TimeUnit;
