+++ skyxplore-ui-test/src/test/java/com/mentoring/pageobject/Page.java
@@ -1,7 +1,7 @@
 package com.mentoring.pageobject;
 
-import com.mentoring.framework.Config;
 import com.mentoring.framework.utils.AllureLogger;
+import com.mentoring.config.Config;
 import io.qameta.allure.Step;
 import lombok.extern.slf4j.Slf4j;
 import org.awaitility.core.ConditionTimeoutException;
