+++ skyxplore-ui-test/src/test/java/com/mentoring/test/CharacterTest.java
@@ -1,7 +1,7 @@
 package com.mentoring.test;
 
 import com.mentoring.framework.BasicTest;
-import com.mentoring.framework.utils.UserUtils;
+import com.mentoring.utilities.UserUtils;
 import com.mentoring.model.User;
 import com.mentoring.pageobject.CharacterSelectionPage;
 import com.mentoring.pageobject.IndexPage;
