+++ skyxplore-ui-test/src/test/java/com/mentoring/test/FriendRequestTest.java
@@ -1,9 +1,9 @@
 package com.mentoring.test;
 
 import com.mentoring.framework.BasicTest;
-import com.mentoring.framework.database.Database;
-import com.mentoring.framework.database.DbSelects;
-import com.mentoring.framework.utils.UserUtils;
+import com.mentoring.database.Database;
+import com.mentoring.database.DbSelects;
+import com.mentoring.utilities.UserUtils;
 import com.mentoring.model.User;
 import com.mentoring.pageobject.*;
 import com.mentoring.test.assertionsteps.CommonAssertions;
