--- skyxplore-ui-test/src/test/resources/skyxplore-localhost-config.properties
@@ -1,6 +1,0 @@
-load_wait = 10
-base_url = http://localhost
-port = 8080
-database_name = sky_xplore_local
-db_user = root
-db_password =
