+++ test-common/src/main/java/com/mentoring/config/Config.java
@@ -1,5 +1,6 @@
-package com.mentoring.utilities;
+package com.mentoring.config;
 
+import com.mentoring.utilities.EmailUtils;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
@@ -8,19 +9,51 @@ import java.util.Objects;
 import java.util.Properties;
 
 @Slf4j
-public final class ConfigUtils {
+public final class Config {
 
+    public static final int LOAD_WAIT;
+    private static final String APPLICATION_URL;
+    public static final Browser BROWSER;
     private static final String SKY_XPLORE_LOCALHOST = "skyxplore-localhost";
     private static final String SKY_XPLORE_PROD = "skyxplore-prod";
     private static final String PATH_WEB_EXTENSION = "/web";
 
+    private static final String BASE_URL;
+    private static final String PORT;
     private static final String ENVIRONMENT = "ENVIRONMENT";
     private static String[] validEnvironmentConfigurationNames = {SKY_XPLORE_LOCALHOST, SKY_XPLORE_PROD};
+    private static final String DATABASE_NAME;
+    private static final String JDBC;
+    private static final String DB_USER;
+    private static final String DB_PASSWORD;
 
-    private ConfigUtils() {
+
+    static {
+        Properties prop = loadProperties();
+        LOAD_WAIT = Integer.parseInt(prop.getProperty("load_wait"));
+        BROWSER = Browser.parse(System.getenv("BROWSER"));
+        PORT = prop.getProperty("port");
+        if (System.getenv(ENVIRONMENT).equalsIgnoreCase(SKY_XPLORE_PROD)) {
+            BASE_URL = EmailUtils.getIpAddressFromEmail();
+        } else {
+            BASE_URL = prop.getProperty("base_url");
+        }
+        APPLICATION_URL = BASE_URL.concat(":").concat(PORT);
+        DATABASE_NAME = prop.getProperty("database_name");
+        if (System.getenv(ENVIRONMENT).equalsIgnoreCase(SKY_XPLORE_LOCALHOST)) {
+            JDBC = "jdbc:mysql://localhost/" + DATABASE_NAME + "?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC";
+        } else {
+            // What should we do on prod? Do we have rights to connect to Prod database?
+            JDBC = "";
+        }
+        DB_USER = prop.getProperty("db_user");
+        DB_PASSWORD = prop.getProperty("db_password");
     }
 
-    public static Properties loadProperties() {
+    private Config() {
+    }
+
+    private static Properties loadProperties() {
         String testEnvironmentName = System.getenv(ENVIRONMENT).toLowerCase();
         String propertiesFileName = validateEnvironment(testEnvironmentName) + "-config.properties";
         return loadProperties(propertiesFileName);
@@ -60,6 +93,26 @@ public final class ConfigUtils {
     }
 
     public static boolean isLocalEnvironmentUsed() {
-        return System.getenv(ConfigUtils.ENVIRONMENT).equals(ConfigUtils.SKY_XPLORE_LOCALHOST);
+        return System.getenv(Config.ENVIRONMENT).equals(Config.SKY_XPLORE_LOCALHOST);
+    }
+
+    public static String getApplicationUrl() {
+        return isLocalEnvironmentUsed() ? APPLICATION_URL + PATH_WEB_EXTENSION : APPLICATION_URL;
+    }
+
+    public static String getBaseUrl() {
+        return APPLICATION_URL;
+    }
+
+    public static String getJBDC() {
+        return JDBC;
+    }
+
+    public static String getDbUser() {
+        return DB_USER;
+    }
+
+    public static String getDbPassword() {
+        return DB_PASSWORD;
     }
 }
