+++ test-common/src/main/java/com/mentoring/exceptions/ConditionNotMetException.java
@@ -1,4 +1,4 @@
-package com.mentoring.framework.exceptions;
+package com.mentoring.exceptions;
 
 public class ConditionNotMetException extends AssertionError {
