+++ test-common/src/main/java/com/mentoring/exceptions/EnvironmentNotSupportedException.java
@@ -1,4 +1,4 @@
-package com.mentoring.framework.exceptions;
+package com.mentoring.exceptions;
 
 import org.testng.SkipException;
