+++ test-common/src/main/java/com/mentoring/utilities/EmailUtils.java
@@ -1,4 +1,4 @@
-package com.mentoring.framework.utils;
+package com.mentoring.utilities;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
