+++ test-common/src/main/java/com/mentoring/utilities/LogUtils.java
@@ -1,4 +1,4 @@
-package com.mentoring.framework.utils;
+package com.mentoring.utilities;
 
 public final class LogUtils {
