+++ test-common/src/main/java/com/mentoring/utilities/RetryStatement.java
@@ -1,4 +1,4 @@
-package com.mentoring.framework.utils;
+package com.mentoring.utilities;
 
 @FunctionalInterface
 public interface RetryStatement {
